Fix bugs. Remove more code.
