+++ admin/src/templates/delete_doc_confirm.html
@@ -1,0 +1,11 @@
+<mm-modal
+  id="'delete-confirm'"
+  danger="true"
+  title-key="'bulkdelete.confirm.title'"
+  submit-key="'bulkdelete.confirm.action'"
+  submitting-key="'Deleting'"
+  on-cancel="cancel()"
+  on-submit="submit()"
+>
+  <p translate>confirm.delete<p>
+</mm-modal>
