+++ admin/src/templates/import_translation.html
